New boss achievements, setbonus achievement fixes
Added 15 boss achievements in accordance with the following trello card: https://trello.com/c/w7F7aTKf/523-miniboss-more-boss-achievements

Added the Legendary Conqueror and Legendary Forger meta-achievements, whcih is a new idea that can hopefully be expanded on in the future

Made several minor fixes to the setbonuses achievements: The unstoppable achivement was removed as the set is no longer in the game, attack and protecc was given bronze instead of copper as its parent achievement, and taking aim was fixed as the achievement required onyx boots instead of silver and a silver chestplate instead of onyx. The "Walking Inferno" achivement was renamed to "Hellwalker".