Fix resource packs not applying
"Modern Pyrotech [1.12] [v1.2].zip" should have no "+" signs as spaces.

Modernity had the wrong version number. I changed it to the version that is present in the 3.1.1a release, however there are a few updates to the resourcepack after 3.6.0 so it might have to be changed if you guys have a newer version in dev.